+++ Changelog.md
@@ -1,4 +1,7 @@
 Change log:
+* 2026-04-28 - U48.1 - 1 new + 2 newly available
+ * Timeless Elf-hawk
+ * Pony of Umbar, Tame Hyena
 * 2026-04-25 - U48 - 1 new + 14 newly available
  * Quintessential Fawn
  * Brown Jerboa, Dusky Parrot, Ginger Mûrai Herding Dog, Overgrown Boglurker, Sunstone Drakeling